Finished day 2

+++ include/aoc.h
@@ -34,8 +34,8 @@ extern "C" {
 int32_t day01_part1(void);
 int32_t day01_part2(void);
 
-int32_t day02_part1(void);
-int32_t day02_part2(void);
+uint64_t day02_part1(void);
+uint64_t day02_part2(void);
 
 #ifdef __cplusplus
 }
